all: rename functions and files
