+++ src/connect/protocol_connect/connect_client.py
@@ -15,7 +15,6 @@ from typing import Any
 import httpcore
 from yarl import URL
 
-from connect.byte_stream import HTTPCoreResponseAsyncByteStream
 from connect.code import Code
 from connect.codec import Codec, StableCodec
 from connect.compression import COMPRESSION_IDENTITY, Compression, get_compresion_from_name
@@ -27,6 +26,7 @@ from connect.connect import (
     StreamType,
     ensure_single,
 )
+from connect.content_stream import BoundAsyncStream
 from connect.error import ConnectError
 from connect.headers import Headers, include_request_headers
 from connect.idempotency_level import IdempotencyLevel
@@ -436,7 +436,7 @@ class ConnectUnaryClientConn(StreamingClientConn):
             hook(response)
 
         assert isinstance(response.stream, AsyncIterable)
-        self.unmarshaler.stream = HTTPCoreResponseAsyncByteStream(response.stream)
+        self.unmarshaler.stream = BoundAsyncStream(response.stream)
 
         await self._validate_response(response)
 
@@ -795,7 +795,7 @@ class ConnectStreamingClientConn(StreamingClientConn):
             hook(response)
 
         assert isinstance(response.stream, AsyncIterable)
-        self.unmarshaler.stream = HTTPCoreResponseAsyncByteStream(aiterator=response.stream)
+        self.unmarshaler.stream = BoundAsyncStream(response.stream)
 
         await self._validate_response(response)
